requirements.txt
cryptography>=42.0.0
pycryptodome>=3.20.0
websockets>=12.0
pydantic>=2.6.0